mainnet/validators/validator1/context/requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
jinja2>=3.0.1
aiofiles>=0.7.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
cryptography>=41.0.5
pyOpenSSL>=23.3.0
websockets>=10.0
redis>=4.2.0

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
aiosqlite>=0.17.0

# Blockchain specific
base58==2.1.1
pycryptodome>=3.19.0

# Monitoring and metrics
prometheus-client>=0.11.0
psutil>=5.9.0
structlog>=21.1.0

# Development dependencies
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910
